+++ ratu/serializers/rfop_serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+
+class RfopSerializer(serializers.Serializer):
+    state = serializers.CharField(max_length=100)
+    kved = serializers.CharField()
+    fullname = serializers.CharField()
+    address = serializers.CharField()
