+++ ratu/urls.py
@@ -1,13 +1,11 @@
 from django.urls import path
-from .views import RfopView
-from .views import RuoView
-from . import views
-from .views import RegionView, DistrictView, CityView, CitydistrictView, StreetView
+from ratu.views.rfop_views import RfopView
+from ratu.views.ruo_views import RuoView
+from ratu.views.ratu_views import RegionView, DistrictView, CityView, CitydistrictView, StreetView
 
 app_name = 'ratu.apps.RatuConfig'
 
 urlpatterns = [
-    path('', views.index, name='index'),
     path('rfop/', RfopView.as_view()),
     path('ruo/', RuoView.as_view()),
     path('region/', RegionView.as_view()),
