+++ ratu/views/ratu_views.py
@@ -2,27 +2,8 @@ from django.shortcuts import render
 from django.http import HttpResponse
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from ratu.models.rfop_models import Rfop
-from ratu.models.ruo_models import Ruo
-from ratu.serializers import RfopSerializer
-from ratu.serializers import RuoSerializer
 from ratu.models.ratu_models import Region, District, City, Citydistrict, Street
-from .serializers import RegionSerializer, CitySerializer, CitydistrictSerializer, StreetSerializer, DistrictSerializer
-
-def index(request):
-    return HttpResponse("Hello, world. You're at the ratu index.")
-
-class RfopView(APIView):
-    def get(self, request):
-        rfop = Rfop.objects.all()[:5]
-        serializer = RfopSerializer(rfop, many=True)
-        return Response({"rfop": serializer.data})
-
-class RuoView(APIView):
-    def get(self, request):
-        ruo = Ruo.objects.all()[:5]
-        serializer = RuoSerializer(ruo, many=True)
-        return Response({"ruo": serializer.data})
+from ratu.serializers.ratu_serializers import RegionSerializer, CitySerializer, CitydistrictSerializer, StreetSerializer, DistrictSerializer
 
 class RegionView(APIView):
     def get(self, request):
